Compendia of the high holidays 2020 -- Release notes
The phone phreaker man named "Lucky" in question uploaded a swastika image into a hacked Google Hangouts chat he initiated with me in 2019-2020 following an email I sent to Christine Blasey Ford on New Years Eve of 2019. The man insisted on being "invited in" (as in a horror flick) to begin that chat conversation with an image about crucifixion. The hateful image was reported by this man to be created in sage-math for an artwork showcase to which he was asked to contribute graphic illustrations. It is in response to one particular set of eyeballs I made in the artwork repository that ended up included in the stock GEyes applet set of eyes as described here: https://github.com/maxieds/GEyesThemes/issues/1. Therefore, the release notes associated with the following link are relevant: https://github.com/maxieds/GEyesThemes/releases/tag/Codename-BackOrifice53. In the email to CBF from 2019, I described the changes to WA state law that happened due to his actions in a physical assault based social engineering ruse perpetrated on a large university campus in Seattle in November of 2014 (also documented in the release notes). I conclude that folklore about 2600 magazine being sponsored by domestic terrorists reported by some conservatives must surely be true. That subversive magazine has had Lucky225 as an editor on it since 2005, the year I took a key programming course in C++ at a large engineering school in the midwest. The origin of the term "phone phreaker" is apparently due to the banned Anarchist Cookbook along with recipes for things like homemade LSD.